Dgoss (#206)
Add dgoss wrapper for goss.
Add getEnv and readFile functions to templates
Add --color flag to force color mode
